Replace ContainerSpec.TypeForwarding with ContainerTests.TypeForwarding.
